requirements.txt
streamlit~=1.40.1
pandas~=2.2.3
pillow~=11.0.0
nltk~=3.9.1
matplotlib~=3.9.2
seaborn~=0.13.2
joblib~=1.4.2
scikit-learn~=1.5.2
imbalanced-learn~=0.12.4
google-api-python-client~=2.153.0
python-dotenv~=1.0.1